Improved File Explorer app.

+++ Apps/FileExplorer/app.py
@@ -27,7 +27,7 @@ def fileexplorer():
 
 	def updatelist():
 		i = 0
-		files = os.listdir("System/Users/Admin/Desktop/")
+		files = os.listdir("System/Users/User/Desktop/")
 		listbox = Listbox(app, selectbackground='SteelBlue', font=("Arial", 10))
 		listbox.place(relx=0, rely=0, relheight=1, relwidth=1)
 		scrollbar = Scrollbar(listbox, orient=VERTICAL, command=listbox.yview)
@@ -38,37 +38,49 @@ def fileexplorer():
 			i += 1
 
 	def copy_file():
-		file_to_copy = fd.askopenfilename(title='Choose a file to copy', filetypes=[("All files", "*.*")], initialdir="System/Users/Admin/Desktop")
-		dir_to_copy_to = fd.askdirectory(title="In which folder to copy to?", initialdir="System/Users/Admin/Desktop")
-		try:
-			shutil.copy(file_to_copy, dir_to_copy_to)
-			mb.showinfo(title='File copied!', message='Your desired file has been copied to your desired location')
-		except:
-			mb.showerror(title='Error!', message='We were unable to copy your file to the desired location. Please try again')
+		file_to_copy = fd.askopenfilename(title='Choose a file to copy', filetypes=[("All files", "*.*")], initialdir="System/Users/User/Desktop")
+		dir_to_copy_to = fd.askdirectory(title="In which folder to copy to?", initialdir="System/Users/User/Desktop")
+		if file_to_copy == "":
+			mb.showerror(title="Error!", message='We\'re not enable to copy this file !')
+		else:
+			if dir_to_copy_to == "":
+				mb.showerror(title="Error!", message='We\'re not enable to copy this file !')
+			else:
+				try:
+					shutil.copy(file_to_copy, dir_to_copy_to)
+					mb.showinfo(title='File copied!', message='Your desired file has been copied to your desired location')
+				except:
+					mb.showerror(title='Error!', message='We were unable to copy your file to the desired location. Please try again')
 	
 	def delete_file():
-		file = fd.askopenfilename(title='Choose a file to delete', filetypes=[("All files", "*.*")], initialdir="System/Users/Admin/Desktop")
-		os.remove(os.path.abspath(file))
-		mb.showinfo('MaxPyOS - File Explorer', message='File has been deleted with success.')
-		i = 0
-		files = os.listdir("System/Users/Admin/Desktop/")
-		listbox = Listbox(app, selectbackground='SteelBlue', font=("Arial", 10))
-		listbox.place(relx=0, rely=0, relheight=1, relwidth=1)
-		scrollbar = Scrollbar(listbox, orient=VERTICAL, command=listbox.yview)
-		scrollbar.pack(side=RIGHT, fill=Y)
-		listbox.config(yscrollcommand=scrollbar.set)
-		while i < len(files):
-			listbox.insert(END, files[i])
-			i += 1
+		file = fd.askopenfilename(title='Choose a file to delete', filetypes=[("All files", "*.*")], initialdir="System/Users/User/Desktop")
+		if file == "":
+			mb.showerror(title="Error!", message='We\'re not enable to delete this file !')
+		else:
+			os.remove(os.path.abspath(file))
+			mb.showinfo('MaxPyOS - File Explorer', message='File has been deleted with success.')
+			i = 0
+			files = os.listdir("System/Users/User/Desktop/")
+			listbox = Listbox(app, selectbackground='SteelBlue', font=("Arial", 10))
+			listbox.place(relx=0, rely=0, relheight=1, relwidth=1)
+			scrollbar = Scrollbar(listbox, orient=VERTICAL, command=listbox.yview)
+			scrollbar.pack(side=RIGHT, fill=Y)
+			listbox.config(yscrollcommand=scrollbar.set)
+			while i < len(files):
+				listbox.insert(END, files[i])
+				i += 1
 
 	def open_file():
-		file = fd.askopenfilename(title='Choose a file of any type', filetypes=[("All files", "*.*")], initialdir="System/Users/Admin/Desktop")
-		split_file = os.path.splitext(file)
-		file_extension = split_file[1]
-		if file_extension == ".txt":
-			openFile(file)
+		file = fd.askopenfilename(title='Choose a file of any type', filetypes=[("All files", "*.*")], initialdir="System/Users/User/Desktop")
+		if file == "":
+			mb.showerror(title="Error!", message='We\'re not enable to open this file !')
 		else:
-			os.startfile(os.path.abspath(file))
+			split_file = os.path.splitext(file)
+			file_extension = split_file[1]
+			if file_extension == ".txt":
+				openFile(file)
+			else:
+				os.startfile(os.path.abspath(file))
 
 	def list_files_in_folder():
 		i = 0
